requirements.txt
discord.py>=2.3.0
python-dateutil>=2.8.0
Flask>=2.3.0